refactor(api): add auth middleware & split hooks
- Add middleware to handle Authorization headers
- Separate SWR hooks into client-hooks.ts
- Upgrade swr from 2.2.5 to 2.3.6
- Update useQuery conditional logic in auth-context
- Remove generate:api build scripts
