+++ package.json
@@ -7,9 +7,9 @@
     "pnpm": ">=9.8.0"
   },
   "scripts": {
-    "postinstall": "husky && pnpm generate:api && pnpm typegen",
-    "dev": "pnpm generate:api && next dev -p 5438",
-    "build": "pnpm generate:api && next build",
+    "postinstall": "husky && pnpm typegen",
+    "dev": "next dev -p 5438",
+    "build": "next build",
     "build:analyze": "ANALYZE=true next build",
     "start": "next start",
     "sitemap": "next-sitemap",
@@ -98,7 +98,7 @@
     "slate-markdown-serializer": "^0.1.5",
     "slate-react": "^0.117.4",
     "sonner": "^2.0.5",
-    "swr": "^2.2.5",
+    "swr": "^2.3.6",
     "swr-openapi": "^5.4.2",
     "tailwind-merge": "^3.3.0",
     "tailwindcss-animate": "^1.0.7",
