+++ shared/api/client-hooks.ts
@@ -1,0 +1,18 @@
+'use client';
+
+import {
+  createQueryHook,
+  createImmutableHook,
+  createInfiniteHook,
+  createMutateHook,
+} from 'swr-openapi';
+
+import { isMatch } from './is-match';
+import { client } from './client';
+
+const PREFIX = 'dao-dao-server-api' as const;
+
+export const useQuery = createQueryHook(client, PREFIX);
+export const useImmutable = createImmutableHook(client, PREFIX);
+export const useInfinite = createInfiniteHook(client, PREFIX);
+export const useMutate = createMutateHook(client, PREFIX, isMatch);
