+++ shared/api/index.ts
@@ -1,11 +1,4 @@
-export {
-  client,
-  useQuery,
-  useImmutable,
-  useInfinite,
-  useMutate,
-  getSwrKey,
-  getSwrKeyWithResponse,
-} from './client';
+export { client, getSwrKey, getSwrKeyWithResponse } from './client';
+export { useQuery, useImmutable, useInfinite, useMutate } from './client-hooks';
 export { ApiError } from './errors';
 export type { paths, components } from './openapi-types';
